First pass at the ThreadPool implementation.
Pools can have two types of children, simple runnables and then consumer
children. We can specify the minimum and maximum number of consumer
children that are allowed and the pool with grow/shrink accordingly. The
algorithm for growing/shrinking could probably use some work. Shrinking is
done when five seconds has passed since that thread was used and we
have more than the minimum.

In addition we have a QueueStrategy that we can use to change how the
queues are handled among consumer children as well as provide custom
behavior for how the queue is selected for enqueing items.

We have a background Watchdog thread that sits in the background and
adjusts the pool size for us. At first I did not have this thread but it
just seemed problematic, especially when we would like to shrink.

If you adjust pool size when things are stopped you can sometimes
sequeeze in a thread that will never get stopped.