Got rid of Tags<>, gonna stick w/ the attribute.
I don't mind it when it's on the same line as the Concern. I actually prefer it.
It's unfortunate that the tag with class benefit will be gone and there's extra noise,
but I think that it's better all in all.

I also plan to introduce a meta type tagging procedure to automate type tagging... we'll see.