Lower level messaging bus infrastructure that I'm building.
This is a simpler, more low level bus than the one that is in MT by
default. It has less of a focus on pub/sub and is more inclined to
point to point operations with pre-configured endpoints. It'd be something
that I build more complex pub/sub scenarios on top of while still
retaining the ability to do return address and avoid pub/sub when
it's unnecessary.

This also has support for collecting a batch of messages in a transport
message. MT has support for batching but it's a little more complicated
than just including them all in one transport message, which is
again a more low level but reasonable way to get the job done. That
transport message also servces as a higher level concept to store
data we'll need, like ReturnAddress, etc...

This is very rough.