Fixed a loophole in behaviors
When a behavior is run, for all protected static fields that are shared between the the context and behavior, the values are copied from the context to the behavior.  The previous implementation was that if a protected static field existed in the behavior and did not exist in the context, that fieldwould not be touched.  In this case, the actual contents of the field would depend on what other contexts had already been run.  If another context that did have a matching field had been previously run, then the value copied for that context would still be in the behavior's field during the execution of the current context (the one not having a matching field).  But if the current context was the first to run, then the field would have its default value.  This usually causes different spec results depending on execution order.

The fix is to require that all static protected fields in the behaviors have matching static protected fields in the contexts so a new value can be copied for every context.  If this isn't true then a SpecificationException describing the problem is thrown.  Conceptually, there should be no reason to have a static protected field in a behavior unless the specs in that behavior intend to use it to receive values from a context, and if that's the case, then the context had better implement a matching source field.  The potential problem here is that if people have already written behaviors that (for some strange reason) have static protected fields that their contexts don't have, and they want it that way, then this would be a breaking change.