Fixing the postbuild step (installutil) of the reporting project
The postbuild step of Machine.Specifications.Reporting only accidentally worked,
apparently because the Windows 7 SDK incorrectly stomps on the $(FrameworkDir)
variable and points it to C:\WINDOWS\Microsoft.NET\Framework\v2.0.50727.  That's
actually not what that variable is supposed to contain, though; it's actually
supposed to contain the path to the *base* folder containing all of the framework
versions.  On machines without the Windows 7 SDK installed, $(FrameworkDir) points
to C:\WINDOWS\Microsoft.NET\Framework or C:\WINDOWS\Microsoft.NET\Framework64 as
it's intended to, but that breaks the postbuild command.  Since we never want to
use the .Net 4.0 version of installutil right now, the correct way to do this
postbuild command is to reference the $(Framework20Dir) variable.  (InstallUtil.exe
is actually contained in the v2.0 framework folder, not v3.5, because v3.5 contains
only additions and differences on top of v2.0.)