Update README.markdown
I straightened out the sections & headings. Installation instructions from the server & nuget. Usage instructions, including patterns & practices, for the major Delegates and Attributes. Retained most of the Command Line Reference documentation from the original. And reformatted and restated the R# and TD.NET integration sections.

I'd like you to turn on the wiki feature of the repository. I'd like to write up the advanced API (behaviors, row-tests, setup per context, assembly-wide control). As well as start re-creating some of the best guidance in wiki pages (or linking to the best articles).

I'll include a CONTRIBUTING.md soon that includes cleaned up instructions on how to build/pull/push and where to join in the community discussions (SO, user groups, etc).