Replaced the Assembly.LoadFile() calls with LoadFrom() in DefaultRunner/Sdk.RunSpecs to fix issue #278
Reason: LoadFrom() uses "Fusion" to resolve the assembly (which will be shadow copied if it is within the base directory and shadow copying is enabled). Previously the LoadFile() calls didn't shadow copy the test assembly in SyncProcessMessage() which lead to a second load of the test assembly within AssemblyContextRunListener.OnAssemblyStart(). Adapted RunSpecs to be symmetric with DefaultRunner.