Introduce a Requires delegate in order to express prerequisites that need to be performed before each Assert, and after the Act
The delegate will typically contain assertions that can be performed within the scope of a unit test.
This will be mostly useful to assert before all tests in a context that the no exceptions were thrown.
Requires delegates can be declared in parent classes when using inheritance in order to apply it to a large number of tests.
In that case, however, the delegates must be declared public.